Split Android review media presentation UI
