+++ apps/web/src/screens/review/components/reviewContentPresentation.ts
@@ -1,7 +1,7 @@
 /**
  * Keep review content presentation heuristics aligned with:
  * - apps/ios/Flashcards/Flashcards/Review/View/ReviewContentPresentation.swift
- * - apps/android/feature/review/src/main/java/com/flashcardsopensourceapp/feature/review/ReviewPresentation.kt
+ * - apps/android/feature/review/src/main/java/com/flashcardsopensourceapp/feature/review/presentation/ReviewContentParser.kt
  */
 export type ReviewContentPresentationMode = "shortPlain" | "paragraphPlain" | "markdown";
